Sends training data to gcp storage.
