+++ requirements.txt
@@ -1,34 +1,75 @@
 attrs==18.2.0
+backcall==0.1.0
+bleach==3.1.0
 cachetools==3.0.0
 certifi==2018.11.29
 chardet==3.0.4
+decorator==4.3.0
+defusedxml==0.5.0
+entrypoints==0.3
 flake8==3.6.0
 flake8-mypy==17.8.0
 google-api-core==1.7.0
 google-auth==1.6.2
 google-auth-oauthlib==0.2.0
+google-cloud==0.34.0
 google-cloud-bigquery==1.8.1
 google-cloud-core==0.29.1
+google-cloud-storage==1.13.2
 google-resumable-media==0.3.2
 googleapis-common-protos==1.5.5
 idna==2.8
+ipykernel==5.1.0
+ipython==7.2.0
+ipython-genutils==0.2.0
+ipywidgets==7.4.2
+jedi==0.13.2
+Jinja2==2.10
+jsonschema==2.6.0
+jupyter==1.0.0
+jupyter-client==5.2.4
+jupyter-console==6.0.0
+jupyter-core==4.4.0
+MarkupSafe==1.1.0
 mccabe==0.6.1
+mistune==0.8.4
 mypy==0.650
 mypy-extensions==0.4.1
+nbconvert==5.4.0
+nbformat==4.4.0
+notebook==5.7.4
 numpy==1.15.4
 oauthlib==2.1.0
 pandas==0.23.4
 pandas-gbq==0.8.0
+pandocfilters==1.4.2
+parso==0.3.1
+pexpect==4.6.0
+pickleshare==0.7.5
+prometheus-client==0.5.0
+prompt-toolkit==2.0.7
 protobuf==3.6.1
+ptyprocess==0.6.0
 pyasn1==0.4.4
 pyasn1-modules==0.2.2
 pycodestyle==2.4.0
 pyflakes==2.0.0
+Pygments==2.3.1
 python-dateutil==2.7.5
 pytz==2018.7
+pyzmq==17.1.2
+qtconsole==4.4.3
 requests==2.21.0
 requests-oauthlib==1.0.0
 rsa==4.0
+Send2Trash==1.5.0
 six==1.12.0
+terminado==0.8.1
+testpath==0.4.2
+tornado==5.1.1
+traitlets==4.3.2
 typed-ast==1.1.1
 urllib3==1.24.1
+wcwidth==0.1.7
+webencodings==0.5.1
+widgetsnbextension==3.4.2
